server/requirements.txt
django
django-ninja
django-cors-headers
jupyter
langchain
langchain-openai
langgraph
langgraph-supervisor
python-decouple
permit
requests
streamlit
psycopg[binary]